Don't rerender on failed sign in,
More specific: Don't rerender during logging in.

Closes #2328

Don't let changes to Accounts._loggingIn invalidate the computation
hasPermission() in main.js. This is done through making Accounts.loggingIn
nonreactive.

This should not prevent the re-run of hasPermission in general, because
whenever Meteor.userId() changes, it's called anyway. It just saves us
from running hasPermissions redundantly and too early (the moment the
user is about to sign in).