chore(security): Invalidate CircleCI caches
We're invalidating the CircleCI caches as part of our response
to the vulnerability introduced through eslint-scope v3.7.2
This is a recommended, proactive, response, though we have no information
to suggest that any of our caches were compromised.

eslint postmortem: https://eslint.org/blog/2018/07/postmortem-for-malicious-package-publishes

The method for invalidating the cache we're using is to add a new ENV var
GLOBAL_CACHE_VERSION which when incremented will invalidate all caches in this
circle.yml file.

This commit also adds the GLOBAL_CACHE_VERSION to our meteor
caches as a separate commit when I discovered that it was not
versioned.