refactor: run Node app in Meteor
This switches up the patterns, such that things like mutations, queries,
GraphQL schemas and resolvers,
startup functions, and other functions are
all injected using the well-known and
highly respected registerPackage function.

This is being done to maintain a comforting
sense of familiarity (and backward compatibility)
amongst the chaos of rewriting, splitting,
and de-Meteorizing a complicated app.

This is a more-or-less working first commit. A few commits will follow
to continue moving around some more files into this new pattern.